move email into background package for reorganization
