+++ enterprise/internal/codemonitors/background/email_templates.go
@@ -1,4 +1,4 @@
-package email
+package background
 
 import (
 	"github.com/sourcegraph/sourcegraph/internal/txemail"
